backend: user search function

+++ scripts/searchfetch.js
@@ -11,6 +11,9 @@ document.addEventListener('DOMContentLoaded', function() {
     } else if (document.getElementById(`lrns-table`)) {
         lrnTableBody = document.querySelector(`#lrns-table tbody`);
         recordType = 'lrn';
+    } else if (document.getElementById(`abstractTiles`)) {
+        recordTableBody = document.getElementById(`abstractTiles`);
+        recordType = 'record';
     }
 
     const tableBodies = {
@@ -21,7 +24,6 @@ document.addEventListener('DOMContentLoaded', function() {
     };
 
     function fetchRecords(recordType, query = '') {
-        console.log(`../../backend/searchfetch.php?record_type=${encodeURIComponent(recordType)}&query=${encodeURIComponent(query)}`); // Debugging line
         
         fetch(`../../backend/searchfetch.php?record_type=${encodeURIComponent(recordType)}&query=${encodeURIComponent(query)}`)
             .then(response => response.json())
@@ -36,19 +38,48 @@ document.addEventListener('DOMContentLoaded', function() {
                     switch (recordType) {
                         case 'record':
                             data.forEach(record => {
-                                const month = Intl.DateTimeFormat('en', { month: 'long' }).format(new Date(record.record_month));
-                                rows += `
-                                    <tr onclick="window.location.href='abstractView.php?abstractId=${record.record_id}';" style="cursor: pointer;">
-                                        <td>${highlightText(record.record_title, query)}</td>
-                                        <td>${highlightText(month + ' ' + record.record_year, query)}</td>
-                                        <td>${highlightText(record.record_authors, query)}</td>
-                                        <td>${highlightText(record.record_trackstrand, query)}</td>
-                                        <td>
-                                            <a href="abstract.php?abstractId=${record.record_id}" class="btn btn-primary btn-sm"><i class='fas fa-edit'></i></a>
-                                            <a href="../../backend/delete.php?abstractId=${record.record_id}" class="btn btn-danger btn-sm delete-button"><i class='fas fa-trash-alt'></i></a>
-                                        </td>
-                                    </tr>
-                                `;
+                                var url = window.location.href;
+                                if (url.includes('admin')) {
+                                    const month = Intl.DateTimeFormat('en', { month: 'long' }).format(new Date(record.record_month));
+                                    rows += `
+                                        <tr onclick="window.location.href='abstractView.php?abstractId=${record.record_id}';" style="cursor: pointer;">
+                                            <td>${highlightText(record.record_title, query)}</td>
+                                            <td>${highlightText(month + ' ' + record.record_year, query)}</td>
+                                            <td>${highlightText(record.record_authors, query)}</td>
+                                            <td>${highlightText(record.record_trackstrand, query)}</td>
+                                            <td>
+                                                <a href="abstract.php?abstractId=${record.record_id}" class="btn btn-primary btn-sm"><i class='fas fa-edit'></i></a>
+                                                <a href="../../backend/delete.php?abstractId=${record.record_id}" class="btn btn-danger btn-sm delete-button"><i class='fas fa-trash-alt'></i></a>
+                                            </td>
+                                        </tr>
+                                    `;
+                                } else {
+                                    rows += `
+                                        <div class="col-12 mb-2">
+                                            <div class="card">
+                                                <div class="card-body" onclick="if (!event.target.closest('button')) { window.location.href='abstractView.php?abstractId=${record.record_id}'; }" style="cursor: pointer;">
+                                                    <div class="row text-center">
+                                                        <div class="col-md-2 d-flex align-items-center justify-content-center border-end">
+                                                            <img src="https://via.placeholder.com/75x100" class="img-fluid rounded-1" alt="${record.record_title}">
+                                                        </div>
+                                                        <div class="col-md-8 d-flex align-items-center justify-content-start border-end">${highlightText(record.record_title, query)}</div>
+                                                        <div class="col-md-2 d-flex align-items-center justify-content-center">
+                                                            <button class="btn btn-outline-primary btn-sm mx-1" data-bs-toggle="modal" data-bs-target="#commentsModal" data-record-id="${record.record_id}" data-record-title="${record.record_title}">
+                                                                <i class="fas fa-comment"></i>
+                                                            </button>
+                                                            <button class="btn btn-outline-danger btn-sm mx-1 like-button" data-record-id="${record.record_id}">
+                                                                <i class="fas fa-heart"></i>
+                                                            </button>
+                                                            <button class="btn btn-outline-success btn-sm mx-1">
+                                                                <i class="fas fa-download"></i>
+                                                            </button>
+                                                        </div>
+                                                    </div>
+                                                </div>
+                                            </div>
+                                        </div>
+                                    `;
+                                }
                             });
                             break;
                         case 'student':
